Make docstrings conform to pydocstyle
This patch fixes the docstrings so that they comply with most of the
pydocstyle rules.

As the development is still much in flux, many public functions lack
documentation so we ignore the corresponding checks.

Pydocstyle forced us to document some of the undocumented functions
which included contracts, so we selectively documented those, while
other public functions of the same modules and/or classes were left
undocumented.