Undry metadata from setup.py
We need to denormalize the medata (such as version, author, license
*etc.*) between `setup.py` and `crosshair/__init__.py` so that both
Sphinx and setup can access them.

Mind that you can not import the module (`crosshair/__init__.py`) from
within `setup.py` as this would cause race condition during the
installation.

To enforce the consistency, we add a script to the pre-commit checks.