Redesign monkeypatching system
We need to intercept functions for three reasons:
(1) to enforce pre/post conditions on PEP316 functions
(2) to replace builtins and library functions with ones that
are more amenable to symbolics.
(3) to short-circuit some function calls, in hopes that
they are not nedded, or their execution can be deferred.

The old mechanism used a "monkey patching" strategy,
which could not reliably perform the replacement and
required us to know all the code we might explore before
running it.
This commit uses a system based on sys.settrace and
hot-swapping functions on the interpreter stack just
before they are invoked.